Consolidate metrics utilities into core module
