--- src/core/metrics/consolidated_metrics.py
@@ -1,148 +1,0 @@
-#!/usr/bin/env python3
-"""
-Consolidated metrics.py - Agent-8 Autonomous Cleanup
-==================================================
-
-Automatically consolidated metrics.py from all scattered locations.
-Generated by Agent-8 during autonomous cleanup mission.
-
-Author: Agent-8 (Integration Enhancement Optimization Manager)
-Mission: Autonomous Cleanup & Side Mission Completion
-Status: MAXIMUM EFFICIENCY MODE
-"""
-
-from typing import Dict, Any, List, Optional
-from dataclasses import dataclass
-from enum import Enum
-import os
-import logging
-
-# ============================================================================
-# TASK MANAGEMENT METRICS
-# ============================================================================
-
-# Consolidated from .\src\core\task_management\unified_scheduler\metrics.py
-# Original file: .\src\core\task_management\unified_scheduler\metrics.py
-# Category: task_management
-# TODO: Import or implement functionality from .\src\core\task_management\unified_scheduler\metrics.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# WORKFLOW METRICS
-# ============================================================================
-
-# Consolidated from .\src\core\workflow\metrics.py
-# Original file: .\src\core\workflow\metrics.py
-# Category: workflow
-# TODO: Import or implement functionality from .\src\core\workflow\metrics.py
-# TODO: Remove duplicate file after migration
-
-# Consolidated from .\src\core\refactoring\workflow_reliability_testing\metrics.py
-# Original file: .\src\core\refactoring\workflow_reliability_testing\metrics.py
-# Category: workflow
-# TODO: Import or implement functionality from .\src\core\refactoring\workflow_reliability_testing\metrics.py
-# TODO: Remove duplicate file after migration
-
-# Consolidated from .\src\workflow\optimization\metrics.py
-# Original file: .\src\workflow\optimization\metrics.py
-# Category: workflow
-# TODO: Import or implement functionality from .\src\workflow\optimization\metrics.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# AGENT MANAGEMENT METRICS
-# ============================================================================
-
-# Consolidated from .\src\core\managers\agents\metrics.py
-# Original file: .\src\core\managers\agents\metrics.py
-# Category: agent_management
-# TODO: Import or implement functionality from .\src\core\managers\agents\metrics.py
-# TODO: Remove duplicate file after migration
-
-# Consolidated from .\src\agents\metrics.py
-# Original file: .\src\agents\metrics.py
-# Category: agent_management
-# TODO: Import or implement functionality from .\src\agents\metrics.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# REFACTORING METRICS
-# ============================================================================
-
-# Consolidated from .\src\core\refactoring\benchmark\metrics.py
-# Original file: .\src\core\refactoring\benchmark\metrics.py
-# Category: refactoring
-# TODO: Import or implement functionality from .\src\core\refactoring\benchmark\metrics.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# OPTIMIZATION METRICS
-# ============================================================================
-
-# Consolidated from .\src\core\optimization\metrics.py
-# Original file: .\src\core\optimization\metrics.py
-# Category: optimization
-# TODO: Import or implement functionality from .\src\core\optimization\metrics.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# BASELINE METRICS
-# ============================================================================
-
-# Consolidated from .\src\core\baseline\metrics.py
-# Original file: .\src\core\baseline\metrics.py
-# Category: baseline
-# TODO: Import or implement functionality from .\src\core\baseline\metrics.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# HANDOFF METRICS
-# ============================================================================
-
-# Consolidated from .\src\core\handoff_reliability\metrics.py
-# Original file: .\src\core\handoff_reliability\metrics.py
-# Category: handoff
-# TODO: Import or implement functionality from .\src\core\handoff_reliability\metrics.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# STATUS MONITORING METRICS
-# ============================================================================
-
-# Consolidated from .\src\services\status_monitor\metrics.py
-# Original file: .\src\services\status_monitor\metrics.py
-# Category: status_monitoring
-# TODO: Import or implement functionality from .\src\services\status_monitor\metrics.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# AI ML METRICS
-# ============================================================================
-
-# Consolidated from .\src\extended\ai_ml\metrics.py
-# Original file: .\src\extended\ai_ml\metrics.py
-# Category: ai_ml
-# TODO: Import or implement functionality from .\src\extended\ai_ml\metrics.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# HEALTH MONITORING METRICS
-# ============================================================================
-
-# Consolidated from .\workspace\health\metrics.py
-# Original file: .\workspace\health\metrics.py
-# Category: health_monitoring
-# TODO: Import or implement functionality from .\workspace\health\metrics.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# CONSOLIDATION METADATA
-# ============================================================================
-
-CONSOLIDATION_INFO = {
-    "pattern": "metrics.py",
-    "total_files": 13,
-    "categories": 10,
-    "consolidated_by": "Agent-8",
-    "mission": "Autonomous Cleanup & Side Mission Completion"
-}
